+++ GEMA/__init__.py
@@ -4,4 +4,4 @@ from .visualization import Visualization
 from .iterativesom import IterativeSOM
 
 name = "GEMA"
-__version__ = "0.4.3"
+__version__ = "0.5.1"